+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:32967/17590f80-13ea-4992-b06c-d81945834b25?"}
+{"mongoUri":"mongodb://127.0.0.1:39807/16dfe5fc-2ab3-42ea-afa6-1d76c8c2b5d9?"}
